move newPanel to common-backend
