+++ extensions/azdata-sanddance/src/extension.ts
@@ -7,13 +7,9 @@ import * as path from 'path';
 import * as vscode from 'vscode';
 import * as azdata from 'azdata';
 import * as tempWrite from 'temp-write';
-import { getWebviewContent } from 'common-backend';
+import { newPanel, WebViewWithUri } from 'common-backend';
 import { MssqlExtensionApi, IFileNode } from './mssqlapis';
 
-interface WebViewWithUri {
-    panel: vscode.WebviewPanel;
-    uriFsPath: string;
-}
 let current: WebViewWithUri | undefined = undefined;
 
 export function activate(context: vscode.ExtensionContext) {
@@ -201,24 +197,3 @@ export function deactivate() {
     vscode.commands.executeCommand('setContext', 'showVisualizer', false);
 }
 
-function newPanel(context: vscode.ExtensionContext, uriFsPath: string, uriTabName?: string) {
-    const webViewWithUri: WebViewWithUri = {
-        panel: vscode.window.createWebviewPanel(
-            'sandDance',
-            `SandDance: ${path.basename(uriTabName || uriFsPath)}`,
-            vscode.ViewColumn.One,
-            {
-                enableScripts: true,
-                // Only allow the webview to access resources in our extension's media directory
-                localResourceRoots: [
-                    vscode.Uri.file(path.join(context.extensionPath, 'resources'))
-                ],
-                retainContextWhenHidden: true
-            }
-        ),
-        uriFsPath
-    };
-    const webView = webViewWithUri.panel.webview;
-    webViewWithUri.panel.webview.html = getWebviewContent(webView, context.extensionPath, uriFsPath);
-    return webViewWithUri;
-}
