+++ extensions/common-backend/src/index.ts
@@ -1,4 +1,4 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-export * from './html';
+export * from './panel';
